extract all top view functions to top_view
